% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_instrument_from_list.R
\name{create_instrument_from_list}
\alias{create_instrument_from_list}
\title{Create instrument from list}
\usage{
create_instrument_from_list(
  question_texts,
  question_numbers = NULL,
  instrument_name = "My instrument"
)
}
\arguments{
\item{question_texts}{A character vector of question texts.}

\item{question_numbers}{A character vector of question numbers. If not provided, the question number will be the index of the question text.}

\item{instrument_name}{A character string of the instrument name.}
}
\description{
This function creates an instrument from a list of questions.
}
\examples{
\donttest{
instrument = create_instrument_from_list(
  list("How old are you?",
    "What is your gender?",
    "What is your name?")
)
}



}
\author{
Alex Nikic
}
