% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_instruments.R
\name{get_example_instruments}
\alias{get_example_instruments}
\title{Retrieve Example Instruments from 'Harmony Data API'}
\usage{
get_example_instruments()
}
\value{
A list representing example instruments retrieved
from the 'Harmony Data API'.
}
\description{
This function retrieves example instruments from the 'Harmony Data API'
using an HTTP POST request.
}
\examples{
\donttest{
# Load required libraries (httr) and call the function
require(httr)
instruments <- get_example_instruments()

# Print the retrieved JSON content
print(instruments)
}

}
\author{
Ulster University [cph]
}
