% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{harmony_stop_for_status}
\alias{harmony_stop_for_status}
\title{Raise an error if the 'Harmony' API returned a failure status}
\usage{
harmony_stop_for_status(res)
}
\arguments{
\item{res}{The response object returned by \code{httr::POST} or \code{httr::GET}.}
}
\value{
No return value, called for side effects.
}
\description{
Converts an unsuccessful HTTP response from the 'Harmony' API into an R error
carrying the \code{detail} message returned by the API, instead of letting the
calling function fail later on an unexpected response body.
}
\author{
Alex Nikic
}
\keyword{internal}
