% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_instruments_from_file.R
\name{tobase64}
\alias{tobase64}
\title{To base64}
\usage{
tobase64(path)
}
\arguments{
\item{path}{Path or link to a file.}
}
\value{
File contents as base64.
}
\description{
Converts a file to base64.
}
