% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_instruments_from_file.R
\name{load_instruments_from_file}
\alias{load_instruments_from_file}
\title{Load Instruments}
\usage{
load_instruments_from_file(path)
}
\arguments{
\item{path}{Path or link to a file.}
}
\value{
List of Instruments.
}
\description{
Open a local file (PDF, Excel, Word or TXT format) and parse it into a list of Instrument objects.
}
