% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_instruments_for_file.R
\name{load_instruments_from_file}
\alias{load_instruments_from_file}
\title{Load Instruments from File}
\usage{
load_instruments_from_file(path)
}
\arguments{
\item{path}{The path to the file to load instruments from.}
}
\value{
A list of instruments returned from the API.
}
\description{
This function loads instruments from a file specified by the \code{path}
parameter and sends the file content to an API for further processing.
It also accepts a URL leading to a file.
}
\examples{
\donttest{
# Load instruments from a PDF file
pdf_file <- "https://www.apa.org/depression-guideline/patient-health-questionnaire.pdf"
response <- load_instruments_from_file(pdf_file)
}

}
\author{
Ulster University [cph]
}
