% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_instruments.R
\name{match_instruments}
\alias{match_instruments}
\title{Match Instruments Function}
\usage{
match_instruments(
  instruments,
  topics = list(),
  is_negate = TRUE,
  clustering_algorithm = "affinity_propagation"
)
}
\arguments{
\item{instruments}{A list of instruments to be matched.}

\item{topics}{A list of topics with which to tag the questions. Default is empty.}

\item{is_negate}{A boolean value to toggle question negation. Default is TRUE.}

\item{clustering_algorithm}{A string value to select the clustering algorithm to use. Must be one of: "affinity_propagation", "kmeans", "deterministic", "hdbscan". Default is "affinity_propagation".}
}
\value{
A list of matched instruments returned from the 'Harmony Data API'.
}
\description{
This function takes a list of instruments, converts it to a format acceptable by the database,
and matches the instruments using the 'Harmony Data API'. It returns the matched instruments.
}
\examples{
\donttest{

instrument_A <- create_instrument_from_list(list(
  "How old are you?",
  "What is your gender?"
))

instrument_B <- create_instrument_from_list(list(
  "Do you smoke?"
))

instruments <- list(instrument_A, instrument_B)

matched_instruments <- match_instruments(
  instruments,
  topics = list("anxiety", "depression")
)
}

}
\author{
Ulster University [cph]
}
